Replace FactoryGirl with Rails Fixtures
Because we are an OAuth heavy application we need real data to test
against. Since we can't use static ids or tokens with FactoryGirl we can
use fixtures.

This also updates the test setup so that we now require two
organizations for testing. One with a free plan and one with a plan for
private repos.

These fixtures cover all variations of assignments and organizations
that this application can have.